% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_specification.R
\name{column_specification}
\alias{column_specification}
\title{Creates column specification for Delta-Neutral EOD file}
\usage{
column_specification(file_type)
}
\description{
Creates column specification for Delta-Neutral EOD file
}
