% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv_option_all.R
\name{csv_option_all}
\alias{csv_option_all}
\title{Reads daily option file from a specific path}
\usage{
csv_option_all(chr_path, exclude_zero_bid = FALSE)
}
\description{
Reads daily option file from a specific path
}
