% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connection.R
\name{db_connection}
\alias{db_connection}
\title{Return a connection object to delta_neutral database}
\usage{
db_connection()
}
\description{
Use this function to get a connection to the delta_neutral database
}
