% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implied_forward.R
\name{implied_forward}
\alias{implied_forward}
\title{Calculate the implied forward price from an option chain}
\usage{
implied_forward(df_full_chain)
}
\description{
Use this function to get calculate the forward price from an option chain.
}
