% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_third_friday.R
\name{is_third_friday}
\alias{is_third_friday}
\title{Determine if a Friday is the third Friday of the month}
\usage{
is_third_friday(dt_friday)
}
\description{
Use this function to determine if a Friday is the third Friday of the month.
}
