% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly_expiration.R
\name{monthly_expiration}
\alias{monthly_expiration}
\title{Regular monthly expiration}
\usage{
monthly_expiration(year, month)
}
\description{
Use this function to get the regular expiration for a given month year.
}
