% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly_last_td.R
\name{monthly_last_td}
\alias{monthly_last_td}
\title{Last trade-date for regular expiration.}
\usage{
monthly_last_td(year, month)
}
\description{
Use this function to get the last trade date for the the regular-expiration
option for this year month.
}
