% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_hist_chain.R
\name{opt_hist_chain}
\alias{opt_hist_chain}
\title{Retrieve all options for a given chain}
\usage{
opt_hist_chain(trade_date, underlying, expiration)
}
\description{
Use this function to get all the option prices for a given trade date,
underlying, expiration combination.
}
