% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option_all.R
\name{option_all}
\alias{option_all}
\title{Retrieve all options for a given trade date.}
\usage{
option_all(trade_date, exclude_zero_bid = FALSE)
}
\description{
Use this function to get all the option prices for a given trade date.
}
\keyword{option}
