% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option_chain.R
\name{option_chain}
\alias{option_chain}
\title{Retrieve all options for a given chain.}
\usage{
option_chain(trade_date, underlying, expiration,
  exclude_zero_bid = FALSE)
}
\description{
Use this function to get all the option prices for a given trade date,
underlying, expiration combination.
}
\keyword{option}
