% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option_underlying.R
\name{option_underlying}
\alias{option_underlying}
\title{Retrieve all options for a given underlying}
\usage{
option_underlying(trade_date, underlying, exclude_zero_bid = FALSE)
}
\description{
Use this function to get all the option prices for a given trade date and
underlying
}
