% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strangle_trade.R
\name{strangle_trade}
\alias{strangle_trade}
\title{Get strangle trades a given underlying, execution date, and expiration}
\usage{
strangle_trade(underlying, execution, expiration, delta_target)
}
\description{
Use this function identify strangle trades for a given chain on a give
execution date
}
