% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap_rate.R
\name{swap_rate}
\alias{swap_rate}
\title{Cacluate the variance swap-rate for an option chain}
\usage{
swap_rate(df_otm, d2x)
}
\description{
Use this function to calculate the variance swap rate for an option chain.
}
