% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/third_friday.R
\name{third_friday}
\alias{third_friday}
\title{Third Friday of month}
\usage{
third_friday(year, month)
}
\description{
Use this function to get the date of the third Friday of the given year and 
month.
}
